rename: quizstore 폴더 이동
